refactor(farmer): split FarmerBundles (526→102) em hook + componentes
Extrai verbatim de src/pages/FarmerBundles.tsx:
- useFarmerBundles: composição dos 3 engines (bundles/argumentos/perguntas
  SPIN) + totais (totalLIE/totalBundles) + expansão de cliente
- CustomerBundleCard, BundleCardFull, QuestionCard, ArgLine, RuleCard,
  types.ts (CustomerCtx), format.ts (fmt)
- +14 testes (helpers puros + presentacionais)

Comportamento preservado 1:1. tsc baseline=0, eslint=0.
